DBZ-7758 Fold abstract processor implementations
